get rid of "csv_tool.py"
